Moved EmojiCollectionViewCell into a separate file. Removed the nib.
