+++ Rocket.Chat/External/RCEmojiKit/Views/EmojiView.swift
@@ -41,62 +41,3 @@ class EmojiView: UIView {
         emojiImageView.frame = self.bounds
     }
 }
-
-class EmojiCollectionViewCell: UICollectionViewCell {
-    enum Emoji {
-        case custom(URL?)
-        case standard(String?)
-    }
-
-    var emojiLabel: UILabel = {
-        let lbl = UILabel()
-        lbl.lineBreakMode = .byClipping
-        lbl.textAlignment = .center
-        lbl.baselineAdjustment = .alignCenters
-        lbl.font = UIFont.systemFont(ofSize: 32)
-        lbl.backgroundColor = UIColor.white
-        return lbl
-    }()
-
-    var emojiImageView: FLAnimatedImageView = {
-        let view = FLAnimatedImageView()
-        view.contentMode = .scaleAspectFit
-        return view
-    }()
-
-    var emoji: Emoji? {
-        didSet {
-            guard let emoji = emoji else {
-                emojiLabel.isHidden = true
-                emojiImageView.isHidden = true
-                return
-            }
-
-            switch emoji {
-            case .custom(let url):
-                emojiImageView.sd_setImage(with: url, completed: nil)
-                emojiImageView.isHidden = false
-            case .standard(let string):
-                emojiLabel.text = string
-                emojiLabel.isHidden = false
-            }
-        }
-    }
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        addSubview(emojiLabel)
-        addSubview(emojiImageView)
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        emoji = nil
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        emojiLabel.frame = self.bounds
-        emojiImageView.frame = self.bounds
-    }
-}
